Shortcuts: None, Desktop, StartMenu

+++ Misc/New-DesktopShortcut.ps1
@@ -7,32 +7,45 @@
         [string]$WorkingDirectory = "", 
         [string]$IconLocation = "", 
         [string]$Arguments = "",
-        [string]$FolderName = "Desktop",
+        [ValidateSet('None','Desktop','StartMenu')]
+        [string]$shortcuts = "Desktop",
         [switch]$RunAsAdministrator = $true
     )
-    $filename = Join-Path ([Environment]::GetFolderPath($FolderName)) "$Name.lnk"
-    if (Test-Path -Path $filename) {
-        Remove-Item $filename -force
-    }
+    if ($shortcuts -ne "None") {
+        
+        if ($shortcuts -eq "Desktop") {
+            $folder = [Environment]::GetFolderPath($shortcuts)
+        } else {
+            $folder = Join-Path ([Environment]::GetFolderPath($shortcuts)) "NavContainerHelper"
+            if (!(Test-Path $folder -PathType Container)) {
+                New-Item $folder -ItemType Directory | Out-Null
+            }
+        }
 
-    $Shell =  New-object -comobject WScript.Shell
-    $Shortcut = $Shell.CreateShortcut($filename)
-    $Shortcut.TargetPath = $TargetPath
-    if (!$WorkingDirectory) {
-        $WorkingDirectory = Split-Path $TargetPath
-    }
-    $Shortcut.WorkingDirectory = $WorkingDirectory
-    if ($Arguments) {
-        $Shortcut.Arguments = $Arguments
-    }
-    if ($IconLocation) {
-        $Shortcut.IconLocation = $IconLocation
-    }
-    $Shortcut.save()
-    if ($RunAsAdministrator) {
-        $bytes = [System.IO.File]::ReadAllBytes($filename)
-        $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
-        [System.IO.File]::WriteAllBytes($filename, $bytes)
+        $filename = Join-Path $folder "$Name.lnk"
+        if (Test-Path -Path $filename) {
+            Remove-Item $filename -force
+        }
+    
+        $Shell =  New-object -comobject WScript.Shell
+        $Shortcut = $Shell.CreateShortcut($filename)
+        $Shortcut.TargetPath = $TargetPath
+        if (!$WorkingDirectory) {
+            $WorkingDirectory = Split-Path $TargetPath
+        }
+        $Shortcut.WorkingDirectory = $WorkingDirectory
+        if ($Arguments) {
+            $Shortcut.Arguments = $Arguments
+        }
+        if ($IconLocation) {
+            $Shortcut.IconLocation = $IconLocation
+        }
+        $Shortcut.save()
+        if ($RunAsAdministrator) {
+            $bytes = [System.IO.File]::ReadAllBytes($filename)
+            $bytes[0x15] = $bytes[0x15] -bor 0x20 #set byte 21 (0x15) bit 6 (0x20) ON
+            [System.IO.File]::WriteAllBytes($filename, $bytes)
+        }
     }
 }
 Export-ModuleMember New-DesktopShortcut
